examples/amazon-bedrock-agentcore/pyproject.toml
[project]
name = "bmasterai-agentcore"
version = "0.1.0"
description = "BMasterAI Research Agent on Amazon Bedrock AgentCore Runtime"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # BMasterAI — structured logging, monitoring, telemetry
    "bmasterai>=0.2.3",

    # Amazon Bedrock AgentCore — managed agent runtime
    "bedrock-agentcore>=0.2.0",
    "bedrock-agentcore-starter-toolkit>=0.2.0",

    # Strands — LLM + tool orchestration framework
    "strands-agents>=0.1.0",

    # AWS SDK
    "boto3>=1.34.0",
    "botocore>=1.34.0",

    # HTTP requests (for fetch_page and search tools)
    "requests>=2.32.0",

    # Local dev helpers
    "python-dotenv>=1.0.0",
]
